Refactor errors to use the same template as non-flat enums
